split tests into test_remote_expert.py and test_moe.py
